move UserData class to newsroom.users
